backend/pyproject.toml
[project]
name = "bidi-demo"
version = "0.1.0"
description = "ADK Gemini Live API Toolkit demo application"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "Google LLC" }
]

dependencies = [
    "google-adk>=1.20.0",
    "aiosqlite>=0.20.0",
    "fastapi>=0.115.0",
    "greenlet>=3.1.1",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
    "google-cloud-aiplatform>=1.133.0",
    "google-cloud-firestore>=2.24.0",
    "google-cloud-storage>=3.8.0",
    "pg8000>=1.31.5",
    "pypdf>=6.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
extend = "../../../pyproject.toml"

[tool.ruff.lint]
ignore = ["C901", "PLR0915"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["app", "bidi_demo"]

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.wheel.shared-data]
"app/static" = "share/bidi-demo/static"
"app/.env" = "share/bidi-demo/.env.template"
